[backend] Rename directory rest -> api
