+++ backend/src/api/contents/video/delete.go
