+++ backend/src/api/rooms/delete.go
