+++ backend/src/api/rooms/get.go
