+++ backend/src/api/rooms/update.go
