+++ backend/src/main.go
@@ -6,13 +6,13 @@ import (
 	"log"
 	"net/http"
 	"os"
+	"secrethome-back/api/contents/video"
+	"secrethome-back/api/rooms"
+	"secrethome-back/api/tags"
 	"secrethome-back/convert"
 	"secrethome-back/features"
 	v1 "secrethome-back/gen/secrethome/v1"
 	"secrethome-back/gen/secrethome/v1/secrethomev1connect"
-	"secrethome-back/rest/contents/video"
-	"secrethome-back/rest/rooms"
-	"secrethome-back/rest/tags"
 
 	"golang.org/x/net/http2"
 	"golang.org/x/net/http2/h2c"
